using buildTypes to simplify setup of legacy admob and make it more clear what's the legacy implementation
