+++ app/src/admob/kotlin/demand/AdMobNative.kt
@@ -4,13 +4,13 @@ import android.view.*
 import androidx.recyclerview.widget.RecyclerView
 import com.adsbynimbus.NimbusAdManager
 import com.adsbynimbus.android.sample.BuildConfig
-import com.adsbynimbus.android.sample.databinding.AdmobNativeAdBinding
+import com.adsbynimbus.android.sample.databinding.AdmobNextGenNativeAdBinding
 import com.adsbynimbus.android.sample.rendering.EmptyAdControllerListenerImplementation
 import com.adsbynimbus.android.sample.rendering.NimbusAdManagerTestListener
 import com.adsbynimbus.render.AdMobRenderer
 import com.adsbynimbus.request.NimbusRequest
 import com.adsbynimbus.request.withAdMobNative
-import com.google.android.gms.ads.nativead.NativeAd
+import com.google.android.libraries.ads.mobile.sdk.nativead.NativeAd
 
 /**
  * These only differ by the type of AdMob NativeAd class.
@@ -19,7 +19,7 @@ object AdMobNative {
     fun show(adManager: NimbusAdManager, adFrame: ViewGroup, logs: RecyclerView, item: String) {
         AdMobRenderer.delegate = object : AdMobRenderer.Delegate {
             override fun customViewForRendering(container: ViewGroup, nativeAd: NativeAd): View {
-                return AdmobNativeAdBinding.inflate(LayoutInflater.from(container.context)).apply {
+                return AdmobNextGenNativeAdBinding.inflate(LayoutInflater.from(container.context)).apply {
                     populateNativeAdView(nativeAd, this)
                 }.root
             }
@@ -42,115 +42,76 @@ object AdMobNative {
         AdMobRenderer.delegate = null
     }
 
-
-    private fun populateNativeAdView(nativeAd: NativeAd, unifiedAdBinding: AdmobNativeAdBinding) {
-        val nativeAdView = unifiedAdBinding.root
-
-        // Set the media view.
-        nativeAdView.mediaView = unifiedAdBinding.adMedia
+    private fun populateNativeAdView(nativeAd: NativeAd, binding: AdmobNextGenNativeAdBinding) {
+        val nativeAdView = binding.root
 
         // Set other ad assets.
-        nativeAdView.headlineView = unifiedAdBinding.adHeadline
-        nativeAdView.bodyView = unifiedAdBinding.adBody
-        nativeAdView.callToActionView = unifiedAdBinding.adCallToAction
-        nativeAdView.iconView = unifiedAdBinding.adAppIcon
-        nativeAdView.priceView = unifiedAdBinding.adPrice
-        nativeAdView.starRatingView = unifiedAdBinding.adStars
-        nativeAdView.storeView = unifiedAdBinding.adStore
-        nativeAdView.advertiserView = unifiedAdBinding.adAdvertiser
+        nativeAdView.headlineView = binding.adHeadline
+        nativeAdView.bodyView = binding.adBody
+        nativeAdView.callToActionView = binding.adCallToAction
+        nativeAdView.iconView = binding.adAppIcon
+        nativeAdView.priceView = binding.adPrice
+        nativeAdView.starRatingView = binding.adStars
+        nativeAdView.storeView = binding.adStore
+        nativeAdView.advertiserView = binding.adAdvertiser
 
         // The headline and media content are guaranteed to be in every UnifiedNativeAd.
-        unifiedAdBinding.adHeadline.text = nativeAd.headline
-        nativeAd.mediaContent?.let { unifiedAdBinding.adMedia.mediaContent = it }
+        binding.adHeadline.text = nativeAd.headline
+        binding.adMedia.mediaContent = nativeAd.mediaContent
 
         // These assets aren't guaranteed to be in every UnifiedNativeAd, so it's important to
         // check before trying to display them.
         if (nativeAd.body == null) {
-            unifiedAdBinding.adBody.visibility = View.INVISIBLE
+            binding.adBody.visibility = View.INVISIBLE
         } else {
-            unifiedAdBinding.adBody.visibility = View.VISIBLE
-            unifiedAdBinding.adBody.text = nativeAd.body
+            binding.adBody.visibility = View.VISIBLE
+            binding.adBody.text = nativeAd.body
         }
 
         if (nativeAd.callToAction == null) {
-            unifiedAdBinding.adCallToAction.visibility = View.INVISIBLE
+            binding.adCallToAction.visibility = View.INVISIBLE
         } else {
-            unifiedAdBinding.adCallToAction.visibility = View.VISIBLE
-            unifiedAdBinding.adCallToAction.text = nativeAd.callToAction
+            binding.adCallToAction.visibility = View.VISIBLE
+            binding.adCallToAction.text = nativeAd.callToAction
         }
 
         if (nativeAd.icon == null) {
-            unifiedAdBinding.adAppIcon.visibility = View.GONE
+            binding.adAppIcon.visibility = View.GONE
         } else {
-            unifiedAdBinding.adAppIcon.setImageDrawable(nativeAd.icon?.drawable)
-            unifiedAdBinding.adAppIcon.visibility = View.VISIBLE
+            binding.adAppIcon.setImageDrawable(nativeAd.icon?.drawable)
+            binding.adAppIcon.visibility = View.VISIBLE
         }
 
         if (nativeAd.price == null) {
-            unifiedAdBinding.adPrice.visibility = View.INVISIBLE
+            binding.adPrice.visibility = View.INVISIBLE
         } else {
-            unifiedAdBinding.adPrice.visibility = View.VISIBLE
-            unifiedAdBinding.adPrice.text = nativeAd.price
+            binding.adPrice.visibility = View.VISIBLE
+            binding.adPrice.text = nativeAd.price
         }
 
         if (nativeAd.store == null) {
-            unifiedAdBinding.adStore.visibility = View.INVISIBLE
+            binding.adStore.visibility = View.INVISIBLE
         } else {
-            unifiedAdBinding.adStore.visibility = View.VISIBLE
-            unifiedAdBinding.adStore.text = nativeAd.store
+            binding.adStore.visibility = View.VISIBLE
+            binding.adStore.text = nativeAd.store
         }
 
         if (nativeAd.starRating == null) {
-            unifiedAdBinding.adStars.visibility = View.INVISIBLE
+            binding.adStars.visibility = View.INVISIBLE
         } else {
-            unifiedAdBinding.adStars.rating = nativeAd.starRating!!.toFloat()
-            unifiedAdBinding.adStars.visibility = View.VISIBLE
-        }
-
-        if (nativeAd.callToAction == null) {
-            unifiedAdBinding.adCallToAction.visibility = View.INVISIBLE
-        } else {
-            unifiedAdBinding.adCallToAction.visibility = View.VISIBLE
-            unifiedAdBinding.adCallToAction.text = nativeAd.callToAction
-        }
-
-        if (nativeAd.icon == null) {
-            unifiedAdBinding.adAppIcon.visibility = View.GONE
-        } else {
-            unifiedAdBinding.adAppIcon.setImageDrawable(nativeAd.icon?.drawable)
-            unifiedAdBinding.adAppIcon.visibility = View.VISIBLE
-        }
-
-        if (nativeAd.price == null) {
-            unifiedAdBinding.adPrice.visibility = View.INVISIBLE
-        } else {
-            unifiedAdBinding.adPrice.visibility = View.VISIBLE
-            unifiedAdBinding.adPrice.text = nativeAd.price
-        }
-
-        if (nativeAd.store == null) {
-            unifiedAdBinding.adStore.visibility = View.INVISIBLE
-        } else {
-            unifiedAdBinding.adStore.visibility = View.VISIBLE
-            unifiedAdBinding.adStore.text = nativeAd.store
-        }
-
-        if (nativeAd.starRating == null) {
-            unifiedAdBinding.adStars.visibility = View.INVISIBLE
-        } else {
-            unifiedAdBinding.adStars.rating = nativeAd.starRating!!.toFloat()
-            unifiedAdBinding.adStars.visibility = View.VISIBLE
+            binding.adStars.rating = nativeAd.starRating!!.toFloat()
+            binding.adStars.visibility = View.VISIBLE
         }
 
         if (nativeAd.advertiser == null) {
-            unifiedAdBinding.adAdvertiser.visibility = View.INVISIBLE
+            binding.adAdvertiser.visibility = View.INVISIBLE
         } else {
-            unifiedAdBinding.adAdvertiser.text = nativeAd.advertiser
-            unifiedAdBinding.adAdvertiser.visibility = View.VISIBLE
+            binding.adAdvertiser.text = nativeAd.advertiser
+            binding.adAdvertiser.visibility = View.VISIBLE
         }
 
         // This method tells the Google Mobile Ads SDK that you have finished populating your
         // native ad view with this native ad.
-        nativeAdView.setNativeAd(nativeAd)
+        nativeAdView.registerNativeAd(nativeAd, binding.adMedia)
     }
 }
