+++ app/src/legacy-admob/kotlin/demand/AdMobInitializer.kt
@@ -1,0 +1,9 @@
+package com.adsbynimbus.android.sample.demand
+
+import android.content.Context
+
+object AdMobInitializer {
+    fun initialize(context: Context) {
+        // No op
+    }
+}
